+++ CHANGELOG.md
@@ -44,6 +44,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-2392 Refine the Account preferences page
 - YLP-2393 Remove fields and "Show more/Show less" logic in the patient header
 - YLP-2394 Hide "Monitoring alerts" card for patients users
+- YLP-2399 Cancel invite dialog new design
 
 ### Engineering use
 - YLP-2044 Remove redux settings store